% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_dot_density}
\alias{as_dot_density}
\title{Convert Census geometry to dots for dot-density mapping}
\usage{
as_dot_density(
  input_data,
  value,
  values_per_dot,
  group = NULL,
  crs = NULL,
  erase_water = FALSE,
  area_threshold = NULL
)
}
\arguments{
\item{input_data}{An input sf object, typically acquired with the tidycensus package}

\item{value}{The value column to be used to determine the number of dots}

\item{values_per_dot}{The number of values per dot; a value of 100 means that each dot will represent approximately 100 values in the value column.}

\item{group}{A column in the dataset that identifies salient groups within which dots should be generated.  For a long-form tidycensus dataset, this will typically be the \code{"variable"} column or some derivative of it.}

\item{crs}{The EPSG code of the coordinate reference system you'd like to use for the output dots. If your data are not already in a projected CRS, supplying this argument will improve performance.}

\item{erase_water}{If \code{TRUE}, calls \code{tigris::erase_water()} to remove water areas from the polygons prior to generating dots. Recommended if your location includes significant water area.}

\item{area_threshold}{The area percentile threshold to be used when erasing water; ranges from 0 (all water area included) to 1 (no water area included)}
}
\value{
The original dataset but of geometry type \code{POINT}, with one point corresponding to the given value:dot ratio for a given group.
}
\description{
Convert Census geometry to dots for dot-density mapping
}
