% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{call_geolocator_latlon}
\alias{call_geolocator_latlon}
\title{Call gelocator for one address with lat/lon}
\usage{
call_geolocator_latlon(lat, lon)
}
\arguments{
\item{lat}{A numeric value}

\item{lon}{A numeric value}
}
\value{
A character string representing the Census block of the supplied
  lat/lon.
}
\description{
Call gelocator for one address with lat/lon
}
\author{
Josie Kressner, \email{josie@transportfoundry.com}
}
