% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{census_api_key}
\alias{census_api_key}
\title{Set the Census API key}
\usage{
census_api_key(api_key)
}
\arguments{
\item{api_key}{A character string representing a user's Census API key.}
}
\description{
Set the Census API key
}
