% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgroups.R
\name{check_ddhca_groups}
\alias{check_ddhca_groups}
\title{Check to see if a given geography / population group combination is available in the Detailed DHC-A file.}
\usage{
check_ddhca_groups(geography, pop_group, state = NULL, county = NULL)
}
\arguments{
\item{geography}{The requested geography.}

\item{pop_group}{The code representing the population group you'd like to check.}

\item{state}{The state (optional)}

\item{county}{The county (optional)}
}
\description{
Check to see if a given geography / population group combination is available in the Detailed DHC-A file.
}
