% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pums.R
\name{df_to_svyrep}
\alias{df_to_svyrep}
\title{Convert a data frame returned by get_pums() to a survey design object}
\usage{
df_to_svyrep(df, type = "person")
}
\arguments{
\item{df}{A data frame with PUMS person or housing replicate weight
variables, most likely returned by \code{\link{get_pums}}.}

\item{type}{Whether to use person or housing-level replicate weights; either
\code{"housing"} or \code{"person"} (the default).}
}
\value{
A svyrep.design object.
}
\description{
This helper function takes a data frame returned by
  \code{\link{get_pums}} and converts it to a svyrep.design object from the
  \code{\link[survey]{svrepdesign}} package. You can then use functions from survey (or
  srvyr) package to calculte weighted estimates with replicate weights
  included to provide accurate standard errors.
}
\examples{
\dontrun{
pums <- get_pums(variables = "AGEP", state = "VT", rep_weights = "person")
pums_design <- df_to_svyrep(pums, type = "person")
survey::svymean(~AGEP, pums_design)
}
}
