% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_census_api_key}
\alias{get_census_api_key}
\title{Check to see if a Census API key is installed}
\usage{
get_census_api_key(key)
}
\description{
Check to see if a Census API key is installed
}
\keyword{internal}
