% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.R
\name{get_cps}
\alias{get_cps}
\title{Load data from the Basic Monthly Current Population Survey Microdata API}
\usage{
get_cps(
  variables,
  state = "all",
  year = NULL,
  month = NULL,
  date_range = NULL,
  var_filter = NULL,
  show_call = FALSE,
  key = NULL
)
}
\arguments{
\item{variables}{A vector of variables from the CPS API. Household and person
IDs, household and person weights, state, year, and month are always
returned.}

\item{state}{A state, or vector of states, for which you would like to
request data. The entire US can be requested with \code{state = "all"}.
Defaults to \code{"all"}.}

\item{year}{The year or years of the CPS data requested.}

\item{month}{The month number or numbers of the CPS data requested.}

\item{date_range}{A vector of two dates (start month and end month). All
months of data between the two dates will be requested You can use either a
date range or a combination of year and month, but not both.}

\item{var_filter}{A named list of CPS variable names and values to be used as
filter when requesting data from the Census API.}

\item{show_call}{If TRUE, display call made to Census API. This can be very
useful in debugging and determining if error messages returned are due to
tidycensus or the Census API. Copy to the API call into a browser and see
what is returned by the API directly. Defaults to FALSE.}

\item{key}{Your Census API key. Obtain one at
\url{http://api.census.gov/data/key_signup.html}.}
}
\value{
A tibble of microdata from the CPS API.
}
\description{
Load data from the Basic Monthly Current Population Survey Microdata API
}
\examples{
\dontrun{
get_cps(variables = "PRTAGE", state = "VT", year = 2020, month = 9)

# All months between July 2018 and and June 2020
# Only return observations where age is between 16 and 24 and
# employment status is 1, 2, or 5
get_cps(
  variables = c("PRTAGE", "PEMLR", "PTDTRACE", "PEHSPNON"),
  state = c("VT", "NH"),
  date_range = c(as.Date("2018-07-01"), as.Date("2020-06-01")),
  var_filter = list(PRTAGE = 16:24, PEMLR = c(1, 2, 5))
  )
}

}
