% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economic.R
\name{get_economic}
\alias{get_economic}
\title{Title}
\usage{
get_economic(geography, variables, naics_code, year = 2012, state = NULL,
  county = NULL, key = NULL, output = "tidy", geometry = FALSE,
  keep_geo_vars = FALSE)
}
\arguments{
\item{keep_geo_vars}{}
}
\description{
Title
}
