% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_codes.R
\docType{data}
\name{mig_recodes}
\alias{mig_recodes}
\title{Dataset with Migration Flows characteristic recodes}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 120 rows and 4 columns.
}
\usage{
data(mig_recodes)
}
\description{
Built-in dataset for Migration Flows code label lookup.

\itemize{
  \item \code{characteristic}: Characteristic variable name
  \item \code{code}: Characteristic calue code
  \item \code{desc}: Characteristic calue label
  \item \code{ordered}: Whether or not recoded value should be ordered factor
}
}
\details{
Dataset with Migration Flows characteristic recodes

Built-in dataset that is created from the
\href{https://www.census.gov/data/developers/data-sets/acs-migration-flows.html}{Migration
Flows API documentation}. This dataset contains labels for the coded values
returned by the Census API and is used when \code{breakdown_labels = TRUE} in
\code{\link{get_flows}}.
}
\keyword{datasets}
