% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe.R
\name{moe_product}
\alias{moe_product}
\title{Calculate the margin of error for a derived product}
\usage{
moe_product(est1, est2, moe1, moe2)
}
\arguments{
\item{est1}{The first factor in the multiplication equation (an estimate)}

\item{est2}{The second factor in the multiplication equation (an estimate)}

\item{moe1}{The margin of error of the first factor}

\item{moe2}{The margin of error of the second factor}
}
\value{
A margin of error for a derived product
}
\description{
Calculate the margin of error for a derived product
}
