% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe.R
\name{moe_prop}
\alias{moe_prop}
\title{Calculate the margin of error for a derived proportion}
\usage{
moe_prop(num, denom, moe_num, moe_denom)
}
\description{
Calculate the margin of error for a derived proportion
}

