% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe.R
\name{moe_sum}
\alias{moe_sum}
\title{Title}
\usage{
moe_sum(...)
}
\arguments{
\item{...}{The margins of error of the variables involved in the sum calculation}
}
\value{
A derived margin of error for a sum
}
\description{
Title
}

