% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_codes.R
\docType{data}
\name{state_laea}
\alias{state_laea}
\title{Dataset of US states with Alaska and Hawaii shifted and re-scaled}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 51 rows and 2 columns.}
\usage{
data(state_laea)
}
\description{
Dataset of US states with Alaska and Hawaii shifted and re-scaled
}
\details{
Built-in dataset for use with the \code{shift_geo} parameter, with the continental United States in a
Lambert azimuthal equal area projection and Alaska and Hawaii shifted and re-scaled.  The data were originally
obtained from the albersusa R package (\url{https://github.com/hrbrmstr/albersusa}).
}
\keyword{datasets}
