% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summary_files}
\alias{summary_files}
\title{Identify summary files for a given decennial Census year}
\usage{
summary_files(year)
}
\arguments{
\item{year}{The year of the decennial Census}
}
\value{
A vector of available summary files for a given decennial Census year. To access data for a given summary file, supply the desired value to the \code{sumfile} parameter in \code{get_decennial()}.
}
\description{
Identify summary files for a given decennial Census year
}
