% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-package.R
\docType{package}
\name{tidycensus-package}
\alias{tidycensus}
\alias{tidycensus-package}
\title{tidycensus: Load US Census Boundary and Attribute Data as 'tidyverse' and 'sf'-Ready Data Frames}
\description{
An integrated R interface to several United States Census Bureau APIs (\url{https://www.census.gov/data/developers/data-sets.html}) and the US Census Bureau's geographic boundary files. Allows R users to return Census and ACS data as tidyverse-ready data frames, and optionally returns a list-column with feature geometry for mapping and spatial analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://walker-data.com/tidycensus/}
  \item Report bugs at \url{https://github.com/walkerke/tidycensus/issues}
}

}
\author{
\strong{Maintainer}: Kyle Walker \email{kyle@walker-data.com}

Authors:
\itemize{
  \item Kyle Walker \email{kyle@walker-data.com}
  \item Matt Herman \email{mfherman@gmail.com}
}

Other contributors:
\itemize{
  \item Kris Eberwein \email{eberwein@knights.ucf.edu} [contributor]
}

}
\keyword{internal}
