% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-package.R
\docType{package}
\name{tidycensus}
\alias{tidycensus-package}
\alias{tidycensus}
\title{Return tidy data frames from the US Census Bureau API}
\description{
This packages uses US Census Bureau data but is neither endorsed nor supported by the US Census Bureau.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://walker-data.com/tidycensus/}
  \item Report bugs at \url{https://github.com/walkerke/tidycensus/issues}
}

}
\author{
Kyle Walker
}
