% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{get_acs}
\alias{get_acs}
\title{Title}
\usage{
get_acs(geography, variables, key, endyear = 2015, output = "tidy",
  state = NULL, county = NULL, geometry = FALSE, keep_geo_vars = FALSE,
  summary_var = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

