% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{get_decennial}
\alias{get_decennial}
\title{Title}
\usage{
get_decennial(geography, variables, year = 2010, sumfile = "sf1",
  state = NULL, county = NULL, geometry = FALSE, output = "tidy",
  keep_geo_vars = FALSE, summary_var = NULL, key = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

