% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{Title}
\usage{
get_estimates(geography, product = NULL, variables = NULL, year = 2017,
  state = NULL, county = NULL, output = "tidy", geometry = FALSE,
  keep_geo_vars = FALSE, shift_geo = FALSE, key = NULL, ...)
}
\arguments{
\item{geography}{geography}

\item{product}{product}

\item{variables}{variables}

\item{year}{year}

\item{state}{state}

\item{county}{county}

\item{output}{output}

\item{geometry}{geometry}

\item{keep_geo_vars}{TRUE or FALSE}

\item{shift_geo}{TRUE or FALSE}

\item{key}{The Census API key}

\item{...}{other keyword arguments}
}
\value{
population estimates
}
\description{
Title
}
