% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe.R
\name{moe_prop}
\alias{moe_prop}
\title{Calculate the margin of error for a derived proportion}
\usage{
moe_prop(num, denom, moe_num, moe_denom)
}
\arguments{
\item{num}{The numerator involved in the proportion calculation (an estimate)}

\item{denom}{The denominator involved in the proportion calculation (an estimate)}

\item{moe_num}{The margin of error of the numerator}

\item{moe_denom}{The margin of error of the denominator}
}
\value{
A margin of error for a derived proportion
}
\description{
Calculate the margin of error for a derived proportion
}
