% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe.R
\name{moe_sum}
\alias{moe_sum}
\title{Calculate the margin of error for a derived sum}
\usage{
moe_sum(...)
}
\arguments{
\item{...}{The margins of error of the variables involved in the sum calculation}
}
\value{
A margin of error for a derived sum
}
\description{
Calculate the margin of error for a derived sum
}

