% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-package.R
\docType{package}
\name{tidycensus}
\alias{tidycensus}
\title{Return tidy data frames from the US Census Bureau API}
\description{
This packages uses US Census Bureau data but is neither endorsed nor supported by the US Census Bureau.
}
\author{
Kyle Walker
}
